colorFactory-chartOptionsBuilder fix
