+++ src/components/visualizations/chart/test/chartOptionsBuilder.spec.ts
@@ -23,7 +23,6 @@ import {
     getBubbleChartSeries,
     getHeatmapDataClasses,
     getTreemapAttributes,
-    isDerivedMeasure,
     getHeatmapSeries,
 } from "../chartOptionsBuilder";
 import { DEFAULT_CATEGORIES_LIMIT } from "../highcharts/commonConfiguration";
@@ -300,35 +299,7 @@ describe("chartOptionsBuilder", () => {
         });
     });
 
-    describe("isDerivedMeasure", () => {
-        it("should return true if measureItem was defined as a popMeasure", () => {
-            const measureItem =
-                fixtures.barChartWithPopMeasureAndViewByAttribute.executionResponse.dimensions[
-                    STACK_BY_DIMENSION_INDEX
-                ].headers[0].measureGroupHeader.items[0];
-            const { afm } = fixtures.barChartWithPopMeasureAndViewByAttribute.executionRequest;
-            expect(isDerivedMeasure(measureItem, afm)).toEqual(true);
-        });
 
-        it("should return true if measureItem was defined as a previousPeriodMeasure", () => {
-            const measureItem =
-                fixtures.barChartWithPreviousPeriodMeasure.executionResponse.dimensions[
-                    STACK_BY_DIMENSION_INDEX
-                ].headers[0].measureGroupHeader.items[0];
-            const { afm } = fixtures.barChartWithPreviousPeriodMeasure.executionRequest;
-            expect(isDerivedMeasure(measureItem, afm)).toEqual(true);
-        });
-
-        it("should return false if measureItem was defined as a simple measure", () => {
-            const measureItem =
-                fixtures.barChartWithPopMeasureAndViewByAttribute.executionResponse.dimensions[
-                    STACK_BY_DIMENSION_INDEX
-                ].headers[0].measureGroupHeader.items[1];
-            const { afm } = fixtures.barChartWithPopMeasureAndViewByAttribute.executionRequest;
-
-            expect(isDerivedMeasure(measureItem, afm)).toEqual(false);
-        });
-    });
 
     describe("getSeriesItemData", () => {
         describe("in usecase of bar chart with pop measure and view by attribute", () => {
